Add files via upload
In this updated script, the "Log-CrashTime" function now accepts a reason parameter, which includes the reason for the crash. The "Monitor-Program" function captures any errors when trying to retrieve the process and logs those errors if the process is not found. This way, the log file will include both the timestamp and the reason for each crash. This, however, will not tell you what mod potentially crashed your server. It will explain other reasonings as to why it crashed.